Move player and mediaitems to seprated packages
