+++ app/src/main/java/com/yurii/youtubemusic/utilities/CategoriesTabAdapter.kt
@@ -2,7 +2,7 @@ package com.yurii.youtubemusic.utilities
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.yurii.youtubemusic.MediaItemsFragment
+import com.yurii.youtubemusic.screens.saved.mediaitems.MediaItemsFragment
 import com.yurii.youtubemusic.models.Category
 
 class CategoriesTabAdapter(fragment: Fragment, private val categories: List<Category>) : FragmentStateAdapter(fragment) {
