+++ app/src/main/java/com/yurii/youtubemusic/utilities/Injector.kt
@@ -8,9 +8,10 @@ import com.yurii.youtubemusic.services.mediaservice.MusicServiceConnection
 import com.yurii.youtubemusic.models.Category
 import com.yurii.youtubemusic.screens.categories.CategoriesEditorViewModel
 import com.yurii.youtubemusic.screens.equalizer.EqualizerViewModel
+import com.yurii.youtubemusic.screens.player.PlayerControllerViewModel
 import com.yurii.youtubemusic.screens.saved.SavedMusicViewModel
+import com.yurii.youtubemusic.screens.saved.mediaitems.MediaItemsViewModel
 import com.yurii.youtubemusic.screens.youtube.YouTubeMusicViewModel
-import com.yurii.youtubemusic.viewmodels.*
 
 object Injector {
 
@@ -24,10 +25,10 @@ object Injector {
         return MediaItemsViewModel(context, category, musicServiceConnection)
     }
 
-    fun providePlayerControllerViewModel(context: Context): PlayerBottomControllerFactory {
+    fun providePlayerControllerViewModel(context: Context): PlayerControllerViewModel.Factory {
         val applicationContext = context.applicationContext
         val musicServiceConnection = provideMusicServiceConnection(context)
-        return PlayerBottomControllerFactory(applicationContext, musicServiceConnection)
+        return PlayerControllerViewModel.Factory(applicationContext, musicServiceConnection)
     }
 
     fun provideSavedMusicViewModel(context: Context): SavedMusicViewModel.Factory {
